+++ CHANGELOG.md
@@ -7,7 +7,8 @@ CHANGELOG
   - Ensure all options using Hash are automatically converted to symbol only hashes
   - Add XLSX option `:freeze` to freeze custom sections of your spreadsheet
   - Add XLSX option `:freeze_headers` to freeze the headers of your spreadsheet
-  - Change Axlsx monkey patch for column autowidth calculations
+  - Remove old Axlsx patch for column width
+  - Backport new code for `string_width` calculations to Axlsx 3.0.1 and below.
 - **3.3.1**
   - [Issue #30](https://github.com/westonganger/spreadsheet_architect/issues/30) - Fix duplicate constant warning for XLSX_COLUMN_TYPES
 - **3.3.0**